# Optimize TNode XML reading